backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
websockets>=12.0

# Google ADK
google-adk>=0.3.0
google-genai>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Browser automation (for UI Navigator)
playwright>=1.40.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
